requirements.txt
anthropic>=0.40.0
gradio>=4.44.0
pypdf>=4.3.0
pandas>=2.2.0
python-dotenv>=1.0.0
